+++ xivsupport/src/main/java/gg/xp/events/XivEntity.java
@@ -2,15 +2,17 @@ package gg.xp.events;
 
 public class XivEntity {
 
-	private final int id;
+	// IMPORTANT: Annoyingly, these all must be 'long' instead of 'int' because the game treats them as
+	// unsigned 32-bit, but Java treats them as signed, so values above 7FFFFFFF cause an overflow
+	private final long id;
 	private final String name;
 
-	public XivEntity(int id, String name) {
+	public XivEntity(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
